requirements.txt
absl-py==0.11.0
amqp==5.0.5
appdirs==1.4.4
asgiref==3.3.1
astor==0.8.1
astunparse==1.6.3
billiard==3.6.3.0
cached-property==1.5.2
cachetools==4.2.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.8
django-allauth==0.44.0
django-filter==2.4.0
dnspython==2.0.0
filelock==3.0.12
flatbuffers==1.12
gast==0.2.2
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.32.0
h5py==2.10.0
html5lib==1.1
httplib2==0.19.0
idna==2.10
importlib-metadata==3.3.0
importlib-resources==4.1.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kombu==5.0.2
Markdown==3.3.3
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
Pillow==8.0.1
pipenv==2020.11.15
prompt-toolkit==3.0.16
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.0
pyparsing==2.4.7
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.3.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
virtualenv==20.2.2
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
